backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
google-generativeai
wikipedia-api
groq
